FORGE-2319: Renamed FacetEventListener to FacetListener and moved to org.jboss.forge.addon.facets package
