+++ facets/api/src/main/java/org/jboss/forge/addon/facets/FacetFactory.java
@@ -7,7 +7,6 @@
 package org.jboss.forge.addon.facets;
 
 import org.jboss.forge.addon.facets.events.FacetEvent;
-import org.jboss.forge.addon.facets.events.FacetEventListener;
 import org.jboss.forge.furnace.spi.ListenerRegistration;
 import org.jboss.forge.furnace.util.Predicate;
 
@@ -126,9 +125,9 @@ public interface FacetFactory
             FACETEDTYPE origin, FACETTYPE facet) throws IllegalArgumentException;
 
    /**
-    * Add a {@link FacetEventListener} to be notified when {@link FacetEvent} events occur.
+    * Add a {@link FacetListener} to be notified when {@link FacetEvent} events occur.
     * 
-    * @see FacetEventListener
+    * @see FacetListener
     */
-   public ListenerRegistration<FacetEventListener> addFacetEventListener(FacetEventListener listener);
+   public ListenerRegistration<FacetListener> addFacetListener(FacetListener listener);
 }
